% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyflights_description.R
\docType{package}
\name{anyflights-description}
\alias{anyflights-description}
\alias{_PACKAGE}
\alias{anyflights-package}
\title{Package: anyflights}
\description{
Query nycflights13-Like Air Travel Data for Specified Years and Airports
}
\details{
The `anyflights` package supplies a set of functions to query air travel
information for specified years and airports. The output of the wrapper
function, `anyflights()`, is a folder of datasets similar to those
found in the data-only package `nycflights13`.
}
\section{anyflights Functions}{

anyflights(), get_airlines(), get_airports(), get_flights(), get_weather()
}

\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/simonpcouch/anyflights}
  \item Report bugs at \url{https://github.com/simonpcouch/anyflights/issues}
}

}
\author{
\strong{Maintainer}: Simon P. Couch \email{simonpatrickcouch@gmail.com}

Authors:
\itemize{
  \item Jay Lee \email{jaylee@reed.edu}
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
  \item Dennis Irorere \email{denironyx@gmail.com} [contributor]
}

}
