% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyflights_description.R
\docType{package}
\name{anyflights_description}
\alias{anyflights_description}
\alias{_PACKAGE}
\alias{anyflights_package}
\title{Package: anyflights}
\description{
Query nycflights13-Like Air Travel Data for Specified Years and Airports
}
\details{
The anyflights package supplies a set of functions to query air travel
information for specified years and airports. The output of the wrapper
function, \code{anyflights()}, is a list of dataframe similar to those
found in the data-only package nycflights13.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/simonpcouch/anyflights}
  \item Report bugs at \url{https://github.com/simonpcouch/anyflights/issues}
}

}
\author{
\strong{Maintainer}: Simon P. Couch \email{simonpatrickcouch@gmail.com}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com}
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
  \item Jay Lee \email{jaylee@reed.edu} [contributor]
  \item Dennis Irorere \email{denironyx@gmail.com} [contributor]
}

}
