% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flights_package.R
\name{as_flights_package}
\alias{as_flights_package}
\title{Generate a Data-Only Package from `anyflights` Data}
\usage{
as_flights_package(data, name)
}
\arguments{
\item{data}{A named list of dataframes outputted by 
\code{\link{anyflights}}.}

\item{name}{The desired name of the resulting package.}

\item{dir}{The folder to write the package to, relative to the current
working directory.}
}
\value{
A directory containing a data-only package built around the
supplied data
}
\description{
Generate a data-only package, including documentation, from data outputted
by the `anyflights` function. Please do not submit the outputted package to
CRAN or similar hosting services!
}
