% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flights.R
\name{make_flights}
\alias{make_flights}
\title{Generate a folder of flights datasets for a given year and airport}
\usage{
make_flights(station, year)
}
\arguments{
\item{station}{The airport of interest (use the airport code)}

\item{year}{The year of interest, as an integer}
}
\value{
A folder named with "airport code" "flights" "year" convention 
(i.e. pdxflights15) containing datasets about flights & relevant metadata
}
\description{
Generate a folder of flights datasets for a given year and airport
}
\examples{
make_flights(station = "PDX", year = 2015)
}
