% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyflights.R
\name{anyflights}
\alias{anyflights}
\title{Generate a folder of flights datasets for a given year and airport}
\usage{
anyflights(station, year, dir)
}
\arguments{
\item{station}{The airport of interest (use the airport code)}

\item{year}{The year of interest, as an integer}

\item{dir}{A character string--the folder for the dataset to be saved in}
}
\value{
A folder containing datasets about air travel
}
\description{
\code{anyflights} is a wrapper function for \code{get_flights},
\code{get_airports}, \code{get_weather}, and \code{get_airlines}
}
\examples{
\donttest{anyflights(station = "PDX", year = 2015, dir = tempdir())}
}
\seealso{
\code{\link{get_flights}} for flight data, 
\code{\link{get_airports}} for airport data, \code{\link{get_weather}} 
for weather data, \code{\link{get_flights}} for flight data, and 
\code{\link{get_airlines}} for airline data
}
