% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyflights_description.R
\docType{package}
\name{anyflights_description}
\alias{anyflights_description}
\alias{anyflights_description-package}
\title{Package: anyflights}
\description{
Download Flight Data For Specified Years and Airports
}
\details{
The `anyflights` package supplies a set of functions to query air travel
information for specified years and airports. The output of the wrapper
function, `anyflights()`, is a folder of datasets similar to those
found in the data-only package `nycflights13`.
}
\section{anyflights Functions}{

anyflights(), get_airlines(), get_airports(), get_flights(), get_weather()
}

