% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyflights_description.R
\docType{package}
\name{anyflights_description}
\alias{anyflights_description}
\alias{anyflights_package}
\alias{anyflights-package}
\title{anyflights: `nycflights13`-Like Data for Specified Years and Airports}
\description{
The anyflights package supplies a set of functions to generate
\code{nycflights13}-like datasets and data packages for specified years and 
airports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/anyflights}
  \item \url{https://simonpcouch.github.io/anyflights/}
  \item Report bugs at \url{https://github.com/simonpcouch/anyflights/issues}
}

}
\author{
\strong{Maintainer}: Simon P. Couch \email{simonpatrickcouch@gmail.com}

Other contributors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com} [contributor]
  \item Jay Lee \email{jaylee@reed.edu} [contributor]
  \item Dennis Irorere \email{denironyx@gmail.com} [contributor]
}

}
