% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_airlines.R
\name{get_airlines}
\alias{get_airlines}
\title{Generate an airlines dataset}
\source{
\url{http://www.transtats.bts.gov/DL_SelectFields.asp?Table_ID=236}
}
\usage{
get_airlines(dir)
}
\arguments{
\item{dir}{A character string---the folder for the dataset to be saved in}
}
\value{
A data frame with ~12 rows and 2 variables:
\describe{
\item{carrier}{Two letter abbreviation}
\item{name}{Full name}
}
}
\description{
Please note that, even with a strong internet connection, this function 
may take several minutes to download relevant data.
}
\examples{
\donttest{get_airlines(dir = tempdir())}
}
\seealso{
\code{\link{get_flights}} for flight data, \code{\link{get_weather}} 
for weather data, \code{\link{get_airports}} for airport data, and 
\code{\link{anyflights}} for a wrapper function
}
