% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_planes.R
\name{get_planes}
\alias{get_planes}
\title{Generate a planes dataset for the specified year}
\format{A data frame with ~3500 rows and 9 variables:
\describe{
\item{tailnum}{Tail number}
\item{year}{Year manufactured}
\item{type}{Type of plane}
\item{manufacturer,model}{Manufacturer and model}
\item{engines,seats}{Number of engines and seats}
\item{speed}{Average cruising speed in mph}
\item{engine}{Type of engine}
}}
\source{
FAA Aircraft registry,
 \url{http://www.faa.gov/licenses_certificates/aircraft_certification/aircraft_registry/releasable_aircraft_download/}
}
\usage{
get_planes(year, dir)
}
\arguments{
\item{year}{The year of interest, as an integer}

\item{dir}{The folder for the dataset to be saved in}
}
\value{
A .Rda dataset of airport data
}
\description{
Generate a planes dataset for the specified year
}
\examples{
get_planes(year = 2015, dir = tempdir())
}
\seealso{
\code{\link{get_flights}} for flight data, \code{\link{get_airports}} for airport
data, \code{\link{get_weather}} for weather data, \code{\link{get_weather}} for airline
data, and \code{\link{make_flights}} for a wrapper function
}
