% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flights.R
\name{make_flights}
\alias{make_flights}
\title{Generate a folder of flights datasets for a given year and airport}
\usage{
make_flights(station, year, dir = tempdir())
}
\arguments{
\item{station}{The airport of interest (use the airport code)}

\item{year}{The year of interest, as an integer}

\item{dir}{The folder for the datasets to be saved in}
}
\value{
A folder containing datasets about air travel
}
\description{
Generate a folder of flights datasets for a given year and airport
}
\examples{
make_flights(station = "PDX", year = 2015)
}
