% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_flights.R
\name{make_flights}
\alias{make_flights}
\title{Generate a folder of flights datasets for a given year and airport}
\usage{
make_flights(station, year, dir = tempdir())
}
\arguments{
\item{station}{The airport of interest (use the airport code)}

\item{year}{The year of interest, as an integer}

\item{dir}{The folder for the datasets to be saved in}
}
\value{
A folder containing datasets about air travel
}
\description{
\code{make_flights} is a wrapper function for \code{get_flights},
\code{get_airports}, \code{get_planes}, \code{get_weather}, and
\code{get_airlines}
}
\examples{
make_flights(station = "PDX", year = 2015)
}
\seealso{
\code{\link{get_flights}} for flight data, \code{\link{get_airports}} for airport
data, \code{\link{get_weather}} for weather data, \code{\link{get_airlines}} for airline
data, and \code{\link{make_planes}} for plane data
}
